lark-skygpt-v3-patched-final (2)/requirements.txt

fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
pydantic>=2.6
pydantic-settings>=2.2
redis>=5.0
sqlalchemy>=2.0
asyncpg>=0.29
openai>=1.40
apscheduler>=3.10
python-multipart>=0.0.9
pdfminer.six>=20231228
pandas>=2.2
openpyxl>=3.1
Pillow>=10.3
tenacity>=8.2
